Move conservatism code to a separate module
