+++ .github/workflows/ci.yml
@@ -18,7 +18,7 @@ jobs:
       fail-fast: false
       matrix:
         gc_plan: [Immix, StickyImmix]
-        moving: [Default, Non_Moving]
+        moving: [Moving, Non_Moving]
     uses: ./.github/workflows/binding-tests.yml
     with:
       gc_plan: ${{ matrix.gc_plan }}
